Add form validation

+++ app/forms.py
@@ -1,9 +1,9 @@
 from flask_wtf import FlaskForm
-from wtforms import StringField, DecimalField, SubmitField
+from wtforms import StringField, FloatField, SubmitField
 from wtforms.validators import DataRequired, Email
 
 
 class DataForm(FlaskForm):
 	email = StringField('Email', validators=[DataRequired(), Email()])
-	height = DecimalField('Height', validators=[DataRequired()])
+	height = FloatField('Height', validators=[DataRequired()])
 	submit = SubmitField('Submit')
